requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
playwright
beautifulsoup4
pypdf
pandas
requests
openai-whisper
